+++ swagger_client/apis/connection_api.py
@@ -2,7 +2,7 @@
 
 """
 ConnectionApi.py
-Copyright 2015 SmartBear Software
+Copyright 2016 SmartBear Software
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -60,25 +60,26 @@ class ConnectionApi(object):
 
         :param callback function: The callback function
             for asynchronous request. (optional)
-        :param int user_id: user_id
-        :param int connector_id: connector_id
-        :param str connect_status: connect_status
-        :param str connect_error: connect_error
-        :param str update_requested_at: update_requested_at
-        :param str update_status: update_status
-        :param str update_error: update_error
-        :param str last_successful_updated_at: last_successful_updated_at
-        :param str created_at: created_at
-        :param str updated_at: updated_at
-        :param int limit: limit
-        :param int offset: offset
-        :param str sort: sort
+        :param str access_token: User's OAuth2 access token
+        :param int user_id: ID of user that owns this correlation
+        :param int connector_id: The id for the connector data source for which the connection is connected
+        :param str connect_status: Indicates whether a connector is currently connected to a service for a user.
+        :param str connect_error: Error message if there is a problem with authorizing this connection.
+        :param str update_requested_at: Time at which an update was requested by a user.
+        :param str update_status: Indicates whether a connector is currently updated.
+        :param str update_error: Indicates if there was an error during the update.
+        :param str last_successful_updated_at: The time at which the connector was last successfully updated.
+        :param str created_at: When the record was first created. Use ISO 8601 datetime format
+        :param str updated_at: When the record was last updated. Use ISO 8601 datetime format
+        :param int limit: The LIMIT is used to limit the number of results returned. So if you have 1000 results, but only want to the first 10, you would set this to 10 and offset to 0. The maximum limit is 200 records.
+        :param int offset: OFFSET says to skip that many rows before beginning to return rows to the client. OFFSET 0 is the same as omitting the OFFSET clause. If both OFFSET and LIMIT appear, then OFFSET rows are skipped before starting to count the LIMIT rows that are returned.
+        :param str sort: Sort by given field. If the field is prefixed with '-', it will sort in descending order.
         :return: InlineResponse2003
                  If the method is called asynchronously,
                  returns the request thread.
         """
 
-        all_params = ['user_id', 'connector_id', 'connect_status', 'connect_error', 'update_requested_at', 'update_status', 'update_error', 'last_successful_updated_at', 'created_at', 'updated_at', 'limit', 'offset', 'sort']
+        all_params = ['access_token', 'user_id', 'connector_id', 'connect_status', 'connect_error', 'update_requested_at', 'update_status', 'update_error', 'last_successful_updated_at', 'created_at', 'updated_at', 'limit', 'offset', 'sort']
         all_params.append('callback')
 
         params = locals()
@@ -91,12 +92,15 @@ class ConnectionApi(object):
             params[key] = val
         del params['kwargs']
 
+
         resource_path = '/connections'.replace('{format}', 'json')
         method = 'GET'
 
         path_params = {}
 
         query_params = {}
+        if 'access_token' in params:
+            query_params['access_token'] = params['access_token']
         if 'user_id' in params:
             query_params['user_id'] = params['user_id']
         if 'connector_id' in params:
@@ -126,7 +130,7 @@ class ConnectionApi(object):
 
         header_params = {}
 
-        form_params = {}
+        form_params = []
         files = {}
 
         body_params = None
@@ -171,13 +175,14 @@ class ConnectionApi(object):
 
         :param callback function: The callback function
             for asynchronous request. (optional)
+        :param str access_token: User's OAuth2 access token
         :param Connection body: Connection that should be stored
-        :return: InlineResponse2004
+        :return: InlineResponse20014
                  If the method is called asynchronously,
                  returns the request thread.
         """
 
-        all_params = ['body']
+        all_params = ['access_token', 'body']
         all_params.append('callback')
 
         params = locals()
@@ -190,16 +195,19 @@ class ConnectionApi(object):
             params[key] = val
         del params['kwargs']
 
+
         resource_path = '/connections'.replace('{format}', 'json')
         method = 'POST'
 
         path_params = {}
 
         query_params = {}
+        if 'access_token' in params:
+            query_params['access_token'] = params['access_token']
 
         header_params = {}
 
-        form_params = {}
+        form_params = []
         files = {}
 
         body_params = None
@@ -226,7 +234,7 @@ class ConnectionApi(object):
                                             body=body_params,
                                             post_params=form_params,
                                             files=files,
-                                            response_type='InlineResponse2004',
+                                            response_type='InlineResponse20014',
                                             auth_settings=auth_settings,
                                             callback=params.get('callback'))
         return response
@@ -247,15 +255,13 @@ class ConnectionApi(object):
         :param callback function: The callback function
             for asynchronous request. (optional)
         :param int id: id of Connection (required)
-        :return: InlineResponse2004
+        :param str access_token: User's OAuth2 access token
+        :return: InlineResponse20014
                  If the method is called asynchronously,
                  returns the request thread.
         """
-        # verify the required parameter 'id' is set
-        if id is None:
-            raise ValueError("Missing the required parameter `id` when calling `connections_id_get`")
 
-        all_params = ['id']
+        all_params = ['id', 'access_token']
         all_params.append('callback')
 
         params = locals()
@@ -268,6 +274,10 @@ class ConnectionApi(object):
             params[key] = val
         del params['kwargs']
 
+        # verify the required parameter 'id' is set
+        if ('id' not in params) or (params['id'] is None):
+            raise ValueError("Missing the required parameter `id` when calling `connections_id_get`")
+
         resource_path = '/connections/{id}'.replace('{format}', 'json')
         method = 'GET'
 
@@ -276,10 +286,12 @@ class ConnectionApi(object):
             path_params['id'] = params['id']
 
         query_params = {}
+        if 'access_token' in params:
+            query_params['access_token'] = params['access_token']
 
         header_params = {}
 
-        form_params = {}
+        form_params = []
         files = {}
 
         body_params = None
@@ -304,7 +316,7 @@ class ConnectionApi(object):
                                             body=body_params,
                                             post_params=form_params,
                                             files=files,
-                                            response_type='InlineResponse2004',
+                                            response_type='InlineResponse20014',
                                             auth_settings=auth_settings,
                                             callback=params.get('callback'))
         return response
@@ -325,16 +337,14 @@ class ConnectionApi(object):
         :param callback function: The callback function
             for asynchronous request. (optional)
         :param int id: id of Connection (required)
+        :param str access_token: User's OAuth2 access token
         :param Connection body: Connection that should be updated
         :return: InlineResponse2002
                  If the method is called asynchronously,
                  returns the request thread.
         """
-        # verify the required parameter 'id' is set
-        if id is None:
-            raise ValueError("Missing the required parameter `id` when calling `connections_id_put`")
 
-        all_params = ['id', 'body']
+        all_params = ['id', 'access_token', 'body']
         all_params.append('callback')
 
         params = locals()
@@ -347,6 +357,10 @@ class ConnectionApi(object):
             params[key] = val
         del params['kwargs']
 
+        # verify the required parameter 'id' is set
+        if ('id' not in params) or (params['id'] is None):
+            raise ValueError("Missing the required parameter `id` when calling `connections_id_put`")
+
         resource_path = '/connections/{id}'.replace('{format}', 'json')
         method = 'PUT'
 
@@ -355,10 +369,12 @@ class ConnectionApi(object):
             path_params['id'] = params['id']
 
         query_params = {}
+        if 'access_token' in params:
+            query_params['access_token'] = params['access_token']
 
         header_params = {}
 
-        form_params = {}
+        form_params = []
         files = {}
 
         body_params = None
@@ -406,15 +422,13 @@ class ConnectionApi(object):
         :param callback function: The callback function
             for asynchronous request. (optional)
         :param int id: id of Connection (required)
+        :param str access_token: User's OAuth2 access token
         :return: InlineResponse2002
                  If the method is called asynchronously,
                  returns the request thread.
         """
-        # verify the required parameter 'id' is set
-        if id is None:
-            raise ValueError("Missing the required parameter `id` when calling `connections_id_delete`")
 
-        all_params = ['id']
+        all_params = ['id', 'access_token']
         all_params.append('callback')
 
         params = locals()
@@ -427,6 +441,10 @@ class ConnectionApi(object):
             params[key] = val
         del params['kwargs']
 
+        # verify the required parameter 'id' is set
+        if ('id' not in params) or (params['id'] is None):
+            raise ValueError("Missing the required parameter `id` when calling `connections_id_delete`")
+
         resource_path = '/connections/{id}'.replace('{format}', 'json')
         method = 'DELETE'
 
@@ -435,10 +453,12 @@ class ConnectionApi(object):
             path_params['id'] = params['id']
 
         query_params = {}
+        if 'access_token' in params:
+            query_params['access_token'] = params['access_token']
 
         header_params = {}
 
-        form_params = {}
+        form_params = []
         files = {}
 
         body_params = None
